refactor(admin): refactor secrets
