+++ apisix/admin/init.lua
@@ -210,6 +210,7 @@ local function run()
         "plugin_configs",
         "consumers",
         "plugin_metadata",
+        "secrets",
     }
     if core.table.array_find(refactored_resources, seg_res) then
         code, data = resource[method](resource, seg_id, req_body, seg_sub_path, uri_args)
